+++ crates/client/data-availability/src/avail/config.rs
@@ -1,6 +1,6 @@
 use serde::Deserialize;
 
-use crate::DaMode;
+use crate::{DaError, DaLayer, DaMode};
 
 const DEFAULT_AVAIL_WS: &str = "ws://127.0.0.1:9945";
 const DEFAULT_APP_ID: u32 = 0;
@@ -19,6 +19,14 @@ pub struct AvailConfig {
     pub seed: String,
     #[serde(default)]
     pub mode: DaMode,
+}
+
+impl TryFrom<&PathBuf> for AvailConfig {
+    type Error = DaErr;
+    fn try_from(path: &PathBuf) -> Result<Self, Self::Error> {
+        let file = File::open(path).map_err(|e| DaError::FailedOpeningConfig(e))?;
+        serde_json::from_reader(file).map_err(|e| DaError::FailedParsingConfig(e))
+    }
 }
 
 fn default_ws() -> String {