+++ crates/node/src/commands/run.rs
@@ -169,7 +169,7 @@ fn init_da_client(da_layer: DaLayer, da_path: &Path) -> Result<Box<dyn DaClient 
         DaLayer::Ethereum => {
             let ethereum_conf: EthereumConfig =
                 serde_json::from_reader(file).map_err(|e| sc_cli::Error::Input(e.to_string()))?;
-            Box::new(EthereumClient::try_from(ethereum_conf)?)
+            Box::new(EthereumClient::try_from(ethereum_conf).map_err(|e| sc_cli::Error::Input(e.to_string()))?)
         }
         #[cfg(feature = "avail")]
         DaLayer::Avail => {